refactor(btc_server): finalize dkg at the end of round 2
