+++ bin/btc-server/src/dkg.rs
@@ -65,6 +65,23 @@ impl App {
                 } else {
                     warn!("Duplicate round2 dkg from peer: {:?}", frost_id);
                 }
+                // If we have a max_signers round2 packages we can generate and save the key package
+                let round2_packages = self.db.get_round2_dkg_packages()?;
+                if round2_packages.len() as u16 == self.max_signers - 1 {
+                    let round1_packages = self.db.get_round1_dkg_packages()?;
+                    if let Some(round2_secret) = self.frost_round2_dkg.lock().unwrap().clone() {
+                        let pk_res = frost::keys::dkg::part3(
+                            &round2_secret,
+                            &round1_packages,
+                            &round2_packages,
+                        )?;
+
+                        self.db.set_key_package(pk_res.0.clone())?;
+                        self.db.set_pubkey_package(pk_res.1.clone())?;
+                        self.db.flush()?;
+                    }
+                }
+
                 return Ok(());
             }
         }
